# Improvement on postings intersection