--- a/index/postings.go
+++ b/index/postings.go
@@ -340,9 +340,9 @@ func (it *intersectPostings) doNext(id uint64) bool {
 
 func (it *intersectPostings) Next() bool {
 	for _, p := range it.arr {
-                if !p.Next() {
-                        return false 
-                }
+		if !p.Next() {
+			return false 
+		}
 		if p.At() > it.cur {
 			it.cur = p.At()
 		}
@@ -352,9 +352,9 @@ func (it *intersectPostings) Next() bool {
 
 func (it *intersectPostings) Seek(id uint64) bool {
 	for _, p := range it.arr {
-                if !p.Seek(id) {
-                        return false 
-                }
+		if !p.Seek(id) {
+			return false 
+		}
 		if p.At() > it.cur {
 			it.cur = p.At()
 		}