Tree-like recursive methods in the current implementation of Intersect() can cause unnecessary Seek().

**An example**
Suppose there are 4 postings [1,2,3,4,5], [3,4, 5], [5,6,7,8], [5, 8].
Calling func Intersect(), p1 and p2 will form an intersectPostings, p3 and p4 will form an intersectPostings. The steps are as follows.
1. Next() for ip1
1.1 Next() for p1 → 1
1.2 Seek(1) for p2 → 3
1.3 Seek(3) for p1 → 3
2. Seek(3) for ip2
2.1 Seek(3) for p3 → 5
2.2 Seek(3) for p4 → 5
3. Seek(5) for ip1
3.1 Seek(5) for p1 → 5
3.2 Seek(5) for p2 → 5
4. Next() for ip1

There are unnecessary Seek(3).
If not use recursive func Intersect(), instead manage all together in an array and Seek() the max value each time. The steps are as follows.
1. Next() for p1, p2, p3, p4 → max=5
2. Seek(5) for p1 → 5
3. Seek(5) for p2 → 5
4. Next() for p1, p2, p3, p4 → finish

**Benchmark**
```
benchmark                       old ns/op     new ns/op     delta
BenchmarkIntersect/case_1-8     199           100           -49.75%
BenchmarkIntersect/case_2-8     195           95.6          -50.97%

benchmark                       old allocs     new allocs     delta
BenchmarkIntersect/case_1-8     4              2              -50.00%
BenchmarkIntersect/case_2-8     4              2              -50.00%

benchmark                       old bytes     new bytes     delta
BenchmarkIntersect/case_1-8     208           96            -53.85%
BenchmarkIntersect/case_2-8     208           96            -53.85%
```