@gouthamve 2019-05-31
After banging my head for a long time, this looks good! I think the recursion causes some overhead in the previous implementation and this new one is actually better in all cases I tested with.