@naivewong 2019-06-01
```
benchmark                       old ns/op     new ns/op     delta
BenchmarkIntersect/case_1-8     210           95.5          -54.52%
BenchmarkIntersect/case_2-8     235           115           -51.06%
BenchmarkIntersect/case_3-8     7617750       385602        -94.94%

benchmark                       old allocs     new allocs     delta
BenchmarkIntersect/case_1-8     4              2              -50.00%
BenchmarkIntersect/case_2-8     4              2              -50.00%
BenchmarkIntersect/case_3-8     99999          1              -100.00%

benchmark                       old bytes     new bytes     delta
BenchmarkIntersect/case_1-8     208           96            -53.85%
BenchmarkIntersect/case_2-8     212           98            -53.77%
BenchmarkIntersect/case_3-8     4799952       32            -100.00%
```