@krasi-georgiev 2019-06-07
much easyer to understand now, Thanks @naivewong 

I found a way to simplify it a bit more.

```
func (it *intersectPostings) doNext(id uint64) bool {
Loop:
	for {
		for _, p := range it.arr {
			if !p.Seek(id) {
				return false
			}
			id = p.At()
			if p.At() > it.cur {
				it.cur = p.At()
				continue Loop
			}
		}
		return true
	}
}

func (it *intersectPostings) Seek(id uint64) bool {
	return it.doNext(id)
}
```

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkIntersect/ManyPostings-8      18495452      649025        -96.49%
BenchmarkIntersect/LongPostings2-8     422           136           -67.77%
BenchmarkIntersect/LongPostings1-8     341           125           -63.34%

benchmark                              old allocs     new allocs     delta
BenchmarkIntersect/ManyPostings-8      99999          1              -100.00%
BenchmarkIntersect/LongPostings1-8     4              2              -50.00%
BenchmarkIntersect/LongPostings2-8     4              2              -50.00%

benchmark                              old bytes     new bytes     delta
BenchmarkIntersect/ManyPostings-8      4799963       32            -100.00%
BenchmarkIntersect/LongPostings2-8     216           98            -54.63%
BenchmarkIntersect/LongPostings1-8     208           96            -53.85%
```